% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{CanESM2.hgt}
\alias{CanESM2.hgt}
\title{Sample data: GCM CanESM2 historical data}
\usage{
data(CanESM2.hgt)
}
\description{
A dataset containing 660 rows (monthly data 1951-2005) and 2496 columns (grids).
}
\keyword{datasets}
