% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{csiro}
\alias{csiro}
\title{Sample data: CSIRO}
\format{
A dataset containing 7 atmospheric variables:
\preformatted{
pr: precipitation (mm day-1)
tas: average surface temperature (deg. C)
dtr: diurnal temperature range (deg. C)
sfcWind: surface wind speed (m s-1)
ps: surface pressure (ps)
huss: surface specific humidity (kg kg-1)
rsds: surface downwelling shortwave radiation (W m-2)
rlds: surface downwelling longwave radiation (W m-2)
}
}
\source{
Cannon, A.J., MBC. \url{data source}
}
\usage{
data(csiro)
}
\description{
Sample data: CSIRO
}
\keyword{datasets}
