% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nc.R
\name{data_cmip}
\alias{data_cmip}
\title{Function: Load NCEP Reanalysis 1 data}
\usage{
data_cmip(
  ncfile,
  var = "psl",
  level,
  start = "1850-01-01",
  by = "month",
  grid,
  mask
)
}
\arguments{
\item{ncfile}{A nc file}

\item{var}{Variable to extract}

\item{level}{Pressure level}

\item{start}{Start Date}

\item{by}{Time step}

\item{grid}{Resample grids, e.g., NCEP reanalysis grids.}

\item{mask}{A shapefile mask, e.g., land mask, ocean mask or country mask.}
}
\value{

}
\description{
Function: Load NCEP Reanalysis 1 data
}
\examples{

}
