% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nc.R
\name{data_ncep}
\alias{data_ncep}
\title{Function: Load NCEP Reanalysis 1 data}
\usage{
data_ncep(ncfile, var = "slp", level, start = "1949-01-01", by = "month", mask)
}
\arguments{
\item{ncfile}{A nc file}

\item{var}{Variable to extract}

\item{level}{Pressure level}

\item{start}{Start Date}

\item{by}{Time step}

\item{mask}{A shapefile mask, land mask, ocean mask or country mask.}
}
\value{

}
\description{
Function: Load NCEP Reanalysis 1 data
}
\examples{

}
