% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nc.R
\name{data_udel}
\alias{data_udel}
\title{Function: Load UDEL Air Temperature & Precipitation}
\usage{
data_udel(
  ncfile,
  var = "precip",
  level,
  start = "1900-01-01",
  by = "month",
  mask
)
}
\arguments{
\item{ncfile}{A nc file}

\item{var}{Variable to extract}

\item{level}{Pressure level}

\item{start}{Start Date}

\item{by}{Time step}

\item{mask}{A shapefile mask, land mask, ocean mask or country mask.}
}
\value{

}
\description{
Function: Load UDEL Air Temperature & Precipitation
}
\examples{

}
