% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{NCEP.uwnd}
\alias{NCEP.uwnd}
\title{Sample data: NCEP Reanalysis 1 data}
\usage{
data(NCEP.uwnd)
}
\description{
A dataset containing 660 rows (monthly data 1951-2005) and 2496 columns (grids).
}
\keyword{datasets}
