% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting_rain.R
\name{nest.rain.cal}
\alias{nest.rain.cal}
\title{Function: Nested bias correction for calibration - rainfall}
\usage{
nest.rain.cal(n, y, obs.sta, start.yr, end.yr, tol = 0.1)
}
\arguments{
\item{n}{The number of columns (e.g., stations)}

\item{y}{A matrix of time series rainfall data}

\item{obs.sta}{Observed (Target) data statistics}

\item{start.yr}{Start year}

\item{end.yr}{End year}

\item{tol}{Tolerance for small values, default 0.1 for rainfall}
}
\value{

}
\description{
Function: Nested bias correction for calibration - rainfall
}
\examples{

}
