% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting_atmos.R
\name{obs.stats}
\alias{obs.stats}
\title{Function: Calculate atmospheric variables statistics}
\usage{
obs.stats(n, y, fun)
}
\arguments{
\item{n}{The number of columns (e.g., stations)}

\item{y}{A matrix of time series atmospheric variables}

\item{fun}{A function to compute the summary statistics, "mean" or "sum".}
}
\value{

}
\description{
Function: Calculate atmospheric variables statistics
}
\examples{

}
