% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting_rain.R
\name{rain.stats}
\alias{rain.stats}
\title{Function: Calculate rainfall statistics}
\usage{
rain.stats(n, y)
}
\arguments{
\item{n}{The number of columns (e.g., stations)}

\item{y}{A matrix of time series rainfall data}
}
\value{

}
\description{
Function: Calculate rainfall statistics
}
\examples{

}
