% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting_atmos.R
\name{rnbc.mod.val}
\alias{rnbc.mod.val}
\title{Function: Recursive Nested bias correction for validation - atmospheric variables}
\usage{
rnbc.mod.val(
  n,
  y,
  obs.sta,
  mod.sta,
  start.yr,
  end.yr,
  fun = "mean",
  tol = 0.1,
  iteration = 3
)
}
\arguments{
\item{n}{The number of columns (e.g., stations)}

\item{y}{A matrix of time series atmospheric variables}

\item{obs.sta}{Observed (Target) data statistics}

\item{mod.sta}{Modeled data statistics at calibration}

\item{start.yr}{Start year}

\item{end.yr}{End year}

\item{fun}{A function to compute the summary statistics, "mean" or "sum".}

\item{tol}{Tolerance for small values, default 0.1}

\item{iteration}{The number of iteration to run the bias correction.}
}
\value{

}
\description{
Function: Recursive Nested bias correction for validation - atmospheric variables
}
\examples{

}
