% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter Plot for statistics check}
\usage{
scatterPlot(obs.sta, mod.sta)
}
\arguments{
\item{obs.sta}{observed statistics}

\item{mod.sta}{modelled statistics}
}
\value{

}
\description{
Scatter Plot for statistics check
}
